fix(contracts): corrigir FK type mismatch nas migrations e presenter/view para index
Migrations (5 arquivos):
- Corrigir tipo de created_by de unsignedBigInteger para unsignedInteger
  em todas as 5 migrations (users.id e int(10) unsigned, nao bigint)
- Corrigir supplier_id e company_id na migration contracts (mesma razao)
- Corrigir asset_id na pivot contract_asset (assets.id e int unsigned)
- Sem essa correcao, php artisan migrate falhava com errno 150
  (Foreign key constraint is incorrectly formed)

Presenters (2 arquivos):
- ContractPresenter::dataTableLayout() retornava array PHP raw
- ContractStatusLabelPresenter::dataTableLayout() mesmo problema
- O componente x-table espera string JSON para data-columns
- Corrigido: return json_encode(layout) seguindo padrao do fork
  (LicensePresenter, AccessoryPresenter, etc.)

Views index (2 arquivos):
- contracts/index.blade.php usava x-table diretamente com name=
- contract-status-labels/index.blade.php mesmo problema
- x-table usa @aware(['name']) que busca de componentes pai,
  nao aceita name como atributo direto
- Corrigido: usar x-table.contracts e x-table.contract-status-labels
  seguindo padrao do fork (x-table.licenses, x-table.accessories, etc.)

Table sub-components (2 arquivos novos):
- blade/table/contracts.blade.php
- blade/table/contract-status-labels.blade.php
- Seguem padrao exato de blade/table/accessories.blade.php

+++ database/migrations/2026_04_12_000001_create_contract_status_labels_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->unsignedInteger('sort_order')->default(0);
             $table->boolean('is_default')->default(false);
             $table->text('notes')->nullable();
-            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedInteger('created_by')->nullable();
             $table->timestamps();
             $table->softDeletes();
